Refactoring variables
Can't provide a default env_user variable as default variable files (i.e. /defaults/main.yml) are not loaded as such if called directly (as opposed to calling the role itself).

There must be a better way to do this (possibly by going back to a standalone role) but this works for now.

Since this is backwards incompatible bumping minor version